feat: rename handleClones to useClones
`useClones` is more-consistent naming with other Vue APIs than `handleClones`.
