+++ docs/.vitepress/config.js
@@ -65,7 +65,7 @@ function getGuideSidebar() {
         { text: 'useFind', link: '/guide/use-find' },
         { text: 'useGet', link: '/guide/use-get' },
         { text: 'usePagination', link: '/guide/use-pagination' },
-        { text: 'handleClones', link: '/guide/handle-clones' },
+        { text: 'useClones', link: '/guide/use-clones' },
       ],
     },
     {
