+++ src/index.ts
@@ -9,12 +9,13 @@ export { clients, registerClient } from './clients'
 
 export { useFind } from './use-find'
 export { useGet } from './use-get'
-export { handleClones } from './handle-clones'
+export { useClones } from './use-clones'
+export { useClones as handleClones } from './use-clones'
 export { usePagination } from './use-pagination'
 
 export { syncWithStorage } from './storage-sync'
 export { syncWithStorageCompressed } from './storage-sync-compressed'
 export { clearStorage } from './clear-storage'
 
-export * from "./types";
-export * from "./service-store/types";
+export * from './types'
+export * from './service-store/types'
